refactor: manual expect
